+++ tests/integration/mod.rs
@@ -1,2 +1,3 @@
+mod crates_alt;
 mod crates_git;
 mod purge_caches;